[cypress][sqllab] remove nested assertions where possible, update to single test index file
