--- superset/assets/cypress/selectors/index.js
@@ -1,1 +1,0 @@
-export const selectResultsTab = () => cy.get('.SouthPane .ReactVirtualized__Table');
